+++ submission/src/epilot_submission/models/submissionpayload.py
@@ -5,8 +5,8 @@ from .optin import OptIn, OptInTypedDict
 from .submissionentity import SubmissionEntity, SubmissionEntityTypedDict
 from epilot_submission.types import BaseModel
 import pydantic
-from typing import List, Optional, TypedDict
-from typing_extensions import Annotated, NotRequired
+from typing import List, Optional
+from typing_extensions import Annotated, NotRequired, TypedDict
 
 
 class SubmissionPayloadTypedDict(TypedDict):